steamslot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
jinja2==3.1.4
python-multipart==0.0.9
cryptography==43.0.3
python-dotenv==1.0.1
